realtime: drop beta wire shape; use single Realtime session encoding
After review on the PR thread, we're dropping the legacy beta representation and the "pick beta vs current" split. Realtime now follows one encoding path for session updates and related client events.
