+++ README.md
@@ -1446,6 +1446,32 @@ final class RealtimeManager {
 }
 ```
 
+#### GA Realtime migration notes
+
+- `realtimeSessionGA(...)` is the GA-native path and supports GA-only `session.update` fields.
+- `realtimeSession(...)` now defaults to GA. Use `apiVersion: .betaV1` only as an explicit opt-in fallback.
+- OpenAI has announced Realtime beta (`OpenAI-Beta: realtime=v1`) deprecation and shutdown on 2026-05-07. Prefer GA paths for new integrations.
+- For `response.create`, GA uses `output_modalities` (not `modalities`).
+- GA `output_modalities` behavior is nuanced:
+  - `["audio"]` returns audio with transcript.
+  - `["text"]` returns text only.
+- For voice mode with built-in web search, use GA tools (`.webSearch`) and GA tool choice.
+
+```swift
+let configuration = OpenAIRealtimeSessionConfigurationGA(
+    outputModalities: [.audio],
+    voice: .builtin("alloy"),
+    tools: [.webSearch(.init(searchContextSize: .medium))],
+    toolChoice: .auto
+)
+
+let session = try await openAIService.realtimeSessionGA(
+    model: "gpt-realtime",
+    configuration: configuration,
+    logLevel: .info
+)
+```
+
 ### How to make a basic request using OpenAI's Responses API
 Note: there is also a streaming version of this snippet below.
