+++ Sources/AIProxy/OpenAI/OpenAIRealtimeConversationItemCreate.swift
@@ -43,7 +43,7 @@ public extension OpenAIRealtimeConversationItemCreate {
             self.role = role
             // The OpenAI Realtime API requires different content types per role:
             //   - user/system: "input_text"
-            //   - assistant: "output_text" (GA-aligned)
+            //   - assistant: "output_text"
             // See: https://platform.openai.com/docs/api-reference/realtime-client-events/conversation/item/create
             let contentType = (role == "assistant") ? "output_text" : "input_text"
             self.content = [Content(type: contentType, text: text)]
